backend/requirements.txt
# FastAPI et serveur
fastapi==0.109.0
uvicorn
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# APIs IA
anthropic==0.18.1
openai==1.12.0
elevenlabs==0.2.27

# Utilitaires
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
pillow==10.2.0

# Monitoring & Logs
loguru==0.7.2
# FastAPI et serveur
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# APIs IA
anthropic==0.18.1
openai==1.12.0
elevenlabs==0.2.27

# Utilitaires
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1
pillow==10.2.0

# Monitoring & Logs
loguru==0.7.2
